Refactor Problem 15 for AOC2024
I had to debug a lot to finally switch to just a single function for
moving items in the warehouse. I kept this commit separate to show the
evolution of how I just used the other move_item function for horizontal
moves the first time I got it to work.